@@ -13,29 +13,32 @@
     ],
     "require": {
         "php": ">=5.6.0",
-        "symfony/framework-bundle": "^2.8|^3.4",
-        "symfony/console": "^2.8|^3.4",
-        "symfony/options-resolver" : "^2.8|^3.4",
-        "symfony/property-access": "^2.8|^3.4",
+        "symfony/framework-bundle": "^3.2|^4",
+        "symfony/console": "^3.2|^4",
+        "symfony/options-resolver" : "^3.2|^4",
+        "symfony/property-access": "^3.2|^4",
         "pagerfanta/pagerfanta": "^1.0.5",
         "psr/log": "^1.0",
         "ruflin/elastica": "^5.2.1"
     },
-    "require-dev":{
-        "doctrine/orm": "^2.4",
+    "require-dev": {
+        "doctrine/orm": "^2.5",
         "doctrine/doctrine-bundle": "^1.6",
-        "doctrine/phpcr-bundle": "^1.2",
-        "doctrine/phpcr-odm": "^1",
-        "jackalope/jackalope-doctrine-dbal": "^1.1",
-        "jms/serializer-bundle": "^2.0",
-        "phpunit/phpunit": "^4.8 || ^5.0",
+        "doctrine/phpcr-bundle": "^1.3",
+        "doctrine/phpcr-odm": "^1.4",
+        "jackalope/jackalope-doctrine-dbal": "^1.2",
+        "jms/serializer-bundle": "^2.2",
+        "phpunit/phpunit": "^5.7.11",
         "propel/propel1": "^1.7",
         "knplabs/knp-components": "^1.2",
-        "symfony/expression-language" : "^2.8|^3.4",
-        "symfony/twig-bundle": "^2.8|^3.4",
-        "symfony/serializer": "^2.8|^3.4",
-        "symfony/yaml": "^2.8|^3.4",
+        "symfony/expression-language" : "^3.2|^4",
+        "symfony/twig-bundle": "^3.2|^4",
+        "symfony/serializer": "^3.2|^4",
+        "symfony/yaml": "^3.2|^4",
         "friendsofphp/php-cs-fixer": "^2.2"
+    },
+    "conflict": {
+        "symfony/dependency-injection": "<3.3"
     },
     "autoload": {
         "psr-4": { "FOS\\ElasticaBundle\\": "" }
